Added retrieval of rapid view by it and sprints associated with a rapid view.

+++ lib/jira_agile_api/resources/estimate_statistic.rb
@@ -6,7 +6,7 @@ module JiraAgileApi
       attribute :client
 
       attribute :statFieldId, String
-      attribute :statFieldValue, Double
+      attribute :statFieldValue, Float
       attribute :statFieldText, String
     end
   end
